+++ src/main/kotlin/com/github/surpsg/diffcoverage/domain/gradle/GradleModule.kt
@@ -1,0 +1,6 @@
+package com.github.surpsg.diffcoverage.domain.gradle
+
+data class GradleModule(
+    val name: String,
+    val absolutePath: String
+)