+++ src/main/kotlin/com/github/surpsg/diffcoverage/extensions/DiffCoverageEngine.kt
@@ -33,7 +33,7 @@ class DiffCoverageEngine : JavaCoverageEngine() {
             getCoverageAnnotator(project) as JavaCoverageAnnotator,
             project,
             suiteBundle,
-            stateBean
+            CoverageViewManager.StateBean()
         )
     }
 }