+++ src/main/kotlin/com/github/surpsg/diffcoverage/services/gradle/GradleService.kt
@@ -1,14 +1,22 @@
 package com.github.surpsg.diffcoverage.services.gradle
 
+import com.github.surpsg.diffcoverage.domain.gradle.GradleTaskWithInitScript
+import com.github.surpsg.diffcoverage.domain.gradle.RunnableGradleTask
+import com.intellij.build.SyncViewManager
+import com.intellij.execution.executors.DefaultRunExecutor
 import com.intellij.openapi.components.Service
-import com.intellij.openapi.externalSystem.action.ExternalSystemActionUtil
-import com.intellij.openapi.externalSystem.model.ProjectSystemId
-import com.intellij.openapi.externalSystem.model.task.TaskData
+import com.intellij.openapi.externalSystem.model.execution.ExternalSystemTaskExecutionSettings
+import com.intellij.openapi.externalSystem.service.execution.ExternalSystemRunConfiguration
 import com.intellij.openapi.externalSystem.service.execution.ProgressExecutionMode
 import com.intellij.openapi.externalSystem.task.TaskCallback
 import com.intellij.openapi.externalSystem.util.ExternalSystemUtil
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.UserDataHolderBase
+import org.jetbrains.plugins.gradle.service.task.GradleTaskManager
 import org.jetbrains.plugins.gradle.settings.GradleSettings
+import org.jetbrains.plugins.gradle.util.GradleConstants
+import kotlin.coroutines.Continuation
+import kotlin.coroutines.suspendCoroutine
 
 @Service
 class GradleService(private val project: Project) {
@@ -17,36 +25,50 @@ class GradleService(private val project: Project) {
         return !GradleSettings.getInstance(project).linkedProjectsSettings.isEmpty()
     }
 
-    fun executeGradleTask(
-        taskName: String,
-        modulePath: String,
-        successCallback: () -> Unit
-    ) {
-        val projectSystemId = ProjectSystemId("GRADLE")
-        val taskExecutionInfo = ExternalSystemActionUtil.buildTaskInfo(
-            TaskData(
-                projectSystemId,
-                taskName,
-                modulePath,
-                ""
+    suspend fun executeTaskWithInitScript(gradleTask: GradleTaskWithInitScript): Boolean {
+        val userData: UserDataHolderBase = UserDataHolderBase().apply {
+            putUserData(
+                GradleTaskManager.INIT_SCRIPT_KEY,
+                gradleTask.initScript
             )
-        )
-        ExternalSystemUtil.runTask(
-            taskExecutionInfo.settings,
-            taskExecutionInfo.executorId,
-            project,
-            projectSystemId,
-            object : TaskCallback {
-                override fun onSuccess() {
-                    successCallback()
-                }
+            putUserData(ExternalSystemRunConfiguration.PROGRESS_LISTENER_KEY, SyncViewManager::class.java)
+        }
+        return suspendCoroutine {
+            executeGradleTask(gradleTask.gradleTask, userData, continuationToTaskCallback(it))
+        }
+    }
 
-                override fun onFailure() {
-                    println("task $taskName failed")
-                }
-            },
+    private fun executeGradleTask(
+        gradleTask: RunnableGradleTask,
+        userData: UserDataHolderBase?,
+        taskCallback: TaskCallback
+    ) {
+        val settings: ExternalSystemTaskExecutionSettings = ExternalSystemTaskExecutionSettings().apply {
+            executionName = gradleTask.taskDescription
+            externalProjectPath = gradleTask.gradleModule.absolutePath
+            taskNames = listOf(gradleTask.taskName)
+            vmOptions = GradleSettings.getInstance(project).gradleVmOptions
+            externalSystemIdString = GradleConstants.SYSTEM_ID.id
+        }
+        ExternalSystemUtil.runTask(
+            settings,
+            DefaultRunExecutor.EXECUTOR_ID,
+            project,
+            GradleConstants.SYSTEM_ID,
+            taskCallback,
             ProgressExecutionMode.IN_BACKGROUND_ASYNC,
-            false
+            false,
+            userData
         )
+    }
+
+    private fun continuationToTaskCallback(continuation: Continuation<Boolean>) = object : TaskCallback {
+        override fun onSuccess() {
+            continuation.resumeWith(Result.success(true))
+        }
+
+        override fun onFailure() {
+            continuation.resumeWith(Result.success(false))
+        }
     }
 }