+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 # diff-coverage-idea-plugin Changelog
 
 ## [Unreleased]
+### Added
+- Keep diff coverage configuration settings per IDE session
+- Added option to set up min coverage in run configuration
+### Changed
+- Diff Coverage running replaced with run configuration instead of shortcut
+- Removed triggering `diffCoverage` Gradle task
+### Fixed
+- Displaying all packages on code coverage view with checked 'Flatten Packages'
+
+## [0.0.4]
 ### Fixed
 - Intellij 2021.1 compatibility
 ### Changed